backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.0
alembic>=1.14.0
httpx>=0.28.0
openai>=1.60.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.7.0
python-multipart>=0.0.18
psycopg2-binary>=2.9.10
